% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_nberblue.R
\name{scale_color_nber}
\alias{scale_color_nber}
\alias{scale_colour_nber}
\alias{scale_fill_nber}
\title{Discrete color scale based on the NBER palette}
\usage{
scale_color_nber(...)

scale_colour_nber(...)

scale_fill_nber(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\value{
None
}
\description{
Discrete color scale based on the NBER palette
}
