% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nber.R
\name{theme_nber}
\alias{theme_nber}
\title{A ggplot theme based on the NBER's Digest and Reporter graphs}
\usage{
theme_nber(base_size = 11, base_family = "")
}
\description{
A ggplot theme based on the NBER's Digest and Reporter graphs
}
